Test 1 v1.2
Changed a lot of things from v1
