+++ Models/DataModel.cs
@@ -1,9 +1,10 @@
 ﻿public class DataModel
 {
-    public int ID { get; set; }
-    public string Name { get; set; }
+    public int ProductID { get; set; }
+    public string ProductName { get; set; }
     public string Brand { get; set; }
     public string Category { get; set; }
+    public string Color { get; set; }
     public decimal Price { get; set; }
 
 }
