+++ Views/DataViewer/TableViewer.cshtml
@@ -25,10 +25,11 @@
 	<table class="table">
 		<thead>
 			<tr>
-				<th>ID</th>
-				<th>Name</th>
+				<th>ProductID</th>
+				<th>ProductName</th>
 				<th>Brand</th>
 				<th>Category</th>
+				<th>Color</th>
 				<th>Price</th>
 			</tr>
 		</thead>
@@ -38,6 +39,7 @@
 				<td>IPhone 15</td>
 				<td>Apple</td>
 				<td>Smartphone</td>
+				<td>Silver</td>
 				<td>3000</td>
 			</tr>
 			<tr>
@@ -45,6 +47,7 @@
 				<td>Samsung Galaxy</td>
 				<td>Samsung</td>
 				<td>Smartphone</td>
+				<td>Blue</td>
 				<td>2000</td>
 			</tr>
 			<tr>
@@ -52,6 +55,7 @@
 				<td>Alienware</td>
 				<td>Dell</td>
 				<td>Laptop</td>
+				<td>Black</td>
 				<td>5000</td>
 			</tr>
 			<tr>
@@ -59,6 +63,7 @@
 				<td>Huawei MatePad</td>
 				<td>Huawei</td>
 				<td>Tablet</td>
+				<td>Black</td>
 				<td>3500</td>
 			</tr>
 			<tr>
@@ -66,6 +71,7 @@
 				<td>Razer Blade 14</td>
 				<td>Razer</td>
 				<td>Laptop</td>
+				<td>Green</td>
 				<td>4500</td>
 			</tr>
 			<tr>
