--- Views/Products/ProductsEditForm.cshtml
@@ -1,32 +1,0 @@
-﻿@model DataToPDFExcel.Models.Products
-
-<head>
-    <title>Products Edit Form</title>
-    <link rel="stylesheet" href="~/css/style.css" />
-</head>
-
-<body>
-    <h2 style="text-align:center;">Edit Product</h2>
-    <div id="form">
-        <form asp-action="EditProduct" method="post">
-            <span>Product ID</span>
-            <input type="text" asp-for="ProductID" readonly />
-            <span>Product Name</span>
-            <input asp-for="ProductName" />
-            <span>Brand</span>
-            <select asp-for="BrandID" asp-items="ViewBag.BrandList" class="form-control">
-                <option value="">-- Select Brand --</option>
-            </select>
-            <span>Category</span>
-            <select asp-for="CategoryID" asp-items="ViewBag.CategoryList" class="form-control">
-                <option value="">-- Select Category --</option>
-            </select>
-            <span>Color</span>
-            <input asp-for="Color" />
-            <span>Price</span>
-            <input asp-for="Price" />
-            <input type="submit" value="Submit" />
-        </form>
-    </div>
-    
-</body>
