--- Views/Products/ProductsEntryForm.cshtml
@@ -1,29 +1,0 @@
-﻿@model DataToPDFExcel.Models.Products
-
-<head>
-    <title>Products Entry Form</title>
-    <link rel="stylesheet" href="~/css/style.css" />
-</head>
-
-<body>
-    <h2 style="text-align:center;">Product Entry</h2>
-    <div id="form">
-        <form asp-action="ProductsEntryForm" method="post">
-            <span>Product Name</span>
-            <input type="text" asp-for="ProductName" />
-            <span>Brand</span>
-            <select asp-for="BrandID" asp-items="ViewBag.BrandList" class="form-control">
-                <option value="">-- Select Brand --</option>
-            </select>
-            <span>Category</span>
-            <select asp-for="CategoryID" asp-items="ViewBag.CategoryList" class="form-control">
-                <option value="">-- Select Category --</option>
-            </select>
-            <span>Color</span>
-            <input type="text" asp-for="Color" />
-            <span>Price</span>
-            <input type="text" asp-for="Price" />
-            <input type="submit" value="Submit">
-        </form>
-    </div>
-</body>
