--- Views/Products/ProductsViewer.cshtml
@@ -1,44 +1,0 @@
-﻿@model List<DataToPDFExcel.Models.ProductsView>
-
-<head>
-    <title>Products Viewer</title>
-    <link rel="stylesheet" href="~/css/style.css" />
-</head>
-
-<body>
-    <h2 style="text-align:center;">Product Summary</h2>
-    
-    <table class="table">
-        <thead>
-            <tr>
-                <th>ID</th>
-                <th>Name</th>
-                <th>Brand</th>
-                <th>Category</th>
-                <th>Color</th>
-                <th>Price</th>
-                <th>Actions</th>
-            </tr>
-        </thead>
-        <tbody>
-            @foreach (var p in Model)
-            {
-                <tr>
-                    <td>@p.ProductID</td>
-                    <td>@p.ProductName</td>
-                    <td>@p.BrandName</td>
-                    <td>@p.CategoryName</td>
-                    <td>@p.Color</td>
-                    <td>@p.Price</td>
-                    <td><a class="btn-link btn-edit" asp-action="ProductsEditForm" asp-route-id="@p.ProductID">Edit</a><a class="btn-link btn-delete" asp-action="DeleteProduct" asp-route-id="@p.ProductID">Delete</a></td>
-                </tr>
-            }
-        </tbody>
-    </table>
-
-    <div style="text-align:right;">
-        <a class="btn-link btn-printexcel" asp-action="ExportToExcel">Export to Excel</a>
-        <a class="btn-link btn-printpdf" asp-action="ExportToPdf">Export to PDF</a>
-    </div>
-
-</body>
