+++ Views/Shared/_Layout.cshtml
@@ -23,10 +23,7 @@
                             <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Index">Home</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-controller="Products" asp-action="ProductsEntryForm">Product Entry</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-controller="Products" asp-action="ProductsViewer">Product Viewer</a>
+                            <a class="nav-link text-dark" asp-controller="DataViewer" asp-action="TableViewer">Product Viewer</a>
                         </li>
                     </ul>
                     <!-- <partial name="_LoginPartial" /> -->
